Consolidate common parameter processing into abstract core class
