+++ core/src/main/java/io/smallrye/openapi/runtime/scanner/spi/ScannerSPILogging.java
@@ -1,0 +1,20 @@
+package io.smallrye.openapi.runtime.scanner.spi;
+
+import org.jboss.logging.Logger;
+import org.jboss.logging.annotations.LogMessage;
+import org.jboss.logging.annotations.Message;
+import org.jboss.logging.annotations.MessageLogger;
+
+@MessageLogger(projectCode = "SROAP", length = 5)
+interface ScannerSPILogging { // NOSONAR (use of constants in an interface)
+    ScannerSPILogging log = Logger.getMessageLogger(ScannerSPILogging.class, ScannerSPILogging.class.getPackage().getName());
+
+    @LogMessage(level = Logger.Level.WARN)
+    @Message(id = 7900, value = "Value '%s' is not a valid %s default")
+    void invalidDefault(String segment, String primitive);
+
+    @LogMessage(level = Logger.Level.WARN)
+    @Message(id = 7901, value = "Matrix parameter references missing path segment: %s")
+    void missingPathSegment(String segment);
+
+}
