+++ extension-spring/src/main/java/io/smallrye/openapi/spring/SpringParameter.java
@@ -3,10 +3,12 @@ package io.smallrye.openapi.spring;
 import org.eclipse.microprofile.openapi.models.parameters.Parameter;
 import org.jboss.jandex.DotName;
 
+import io.smallrye.openapi.runtime.scanner.spi.FrameworkParameter;
+
 /**
  * Meta information for the Spring Parameter annotations relating them
  * to the In and Style attributes of Parameters.
- * 
+ *
  * @author Phillip Kruger (phillip.kruger@redhat.com)
  */
 public enum SpringParameter {
@@ -19,40 +21,27 @@ public enum SpringParameter {
     //BEAN_PARAM(SpringConstants.BEAN_PARAM, null, null, null),
     //FORM_PARAM(SpringConstants.FORM_PARAM, null, Parameter.Style.FORM, Parameter.Style.FORM),
 
-    private final DotName name;
-    final Parameter.In location;
-    final Parameter.Style style;
-    final Parameter.Style defaultStyle;
-    final String mediaType;
+    final FrameworkParameter parameter;
 
     private SpringParameter(DotName name, Parameter.In location, Parameter.Style style, Parameter.Style defaultStyle,
             String mediaType) {
-        this.name = name;
-        this.location = location;
-        this.style = style;
-        this.defaultStyle = defaultStyle;
-        this.mediaType = mediaType;
+        this.parameter = new FrameworkParameter(name, location, style, defaultStyle, mediaType);
     }
 
     private SpringParameter(DotName name, Parameter.In location, Parameter.Style style, Parameter.Style defaultStyle) {
         this(name, location, style, defaultStyle, null);
     }
 
-    static SpringParameter forName(DotName annotationName) {
+    static FrameworkParameter forName(DotName annotationName) {
         for (SpringParameter value : values()) {
-            if (value.name.equals(annotationName)) {
-                return value;
+            if (value.parameter.getName().equals(annotationName)) {
+                return value.parameter;
             }
         }
         return null;
     }
 
     public static boolean isParameter(DotName annotationName) {
-        for (SpringParameter value : values()) {
-            if (value.name.equals(annotationName)) {
-                return true;
-            }
-        }
-        return false;
+        return forName(annotationName) != null;
     }
 }
