+++ extension-vertx/src/main/java/io/smallrye/openapi/vertx/VertxAnnotationScanner.java
@@ -273,7 +273,7 @@ public class VertxAnnotationScanner extends AbstractAnnotationScanner {
             PathItem pathItem = new PathItemImpl();
             Function<AnnotationInstance, Parameter> reader = t -> ParameterReader.readParameter(context, t);
 
-            ResourceParameters params = ParameterProcessor.process(context.getIndex(), context.getClassLoader(), resourceClass,
+            ResourceParameters params = ParameterProcessor.process(context, resourceClass,
                     method, reader,
                     context.getExtensions());
             operation.setParameters(params.getOperationParameters());
